LadderGame의 검증 로직 이동 -refactor: Validator class 새로 추가 -test: ValidatorTest 작성
